backend/requirements.txt
django>=4.2.7
djangorestframework>=3.14.0
psycopg2-binary>=2.9.9
django-cors-headers>=4.3.0
channels>=4.0.0
channels-redis>=4.1.0
celery>=5.3.4
redis>=5.0.1
openai>=1.3.5
python-multipart>=0.0.6
pillow>=10.1.0
python-dotenv>=1.0.0
